==> R/1. Setup.Rmd <==
---
title: "1. Setup"
author: "J Hagy"
date: "2023-12-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

## Check if packages are installed, and if not, install them

```{r}

# The Delaware DO Repository Programs use the following R Packages
packages <- c("betareg","EGRET","gratia","gt","here","janitor","knitr",
              "lmodel2","marelac","mgcv","modelr","patchwork",
              "qacBase","qgam","quantreg","readxl","stats","tidyverse",
              "viridis","wesanderson","wql","zoo")   

df <- packages %in% installed.packages()

if (all(df)) {
  print("All the packages needed to run the R programs in the DelawareDO repository are installed")
} else {
  print("The following required packages are not installed and should be installed using install.packages():\n")
  print(packages[!df])
}


```
