==> R/5. Assimilate Chester WQ Time Series Data.Rmd <==
---
title: "5. Assimilate Chester WQ Time Series Data"
author: "J. Hagy"
date: "`r Sys.Date()`"
output:
  html_document:
    df_print: paged
editor_options: 
  chunk_output_type: console
---

This script reads the continuous water quality time series data for the Delaware River at Chester, PA from the USGS water quality portal and tidys the data for use in this analysis.  

This script also uses the wql package to calculate salinity from specific conductivity and the marelac package to calculate percent oxygen saturation from dissolved oxygen concentration (mg/L).

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#install.packages("wql")
#https://rdrr.io/cran/wql/
# Exploring Water Quality Monitoring Data - Jassby, Cloern, Stachelek

library(tidyverse)
library(lubridate)
library(marelac)
library(wql)
library(wesanderson)
library(viridis)
library(here)

large_text <-
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14))
no_grid <- theme(panel.grid=element_blank())

pal <- wes_palette("Zissou1",5,type="discrete")

# Get function to impute missing values
source(here("functions","imputeWQbyYear.R"))

```

## Download Chester WQ Data
This code chunk creates the data frame drChester_ctdo which has variables for daily mean conductivity (µS/cm), water temperature (°C), and dissolved oxygen (mg/L) for 1961-12-27 through 20213-1-30.  An Rdata file was saved when the code was originally created but is not overwritten to preserve data integrity.

```{r download data}

# URL generated by NWIS for later use.
url <- "https://waterdata.usgs.gov/nwis/dv?cb_00010=on&cb_00011=on&cb_00095=on&cb_00300=on&format=rdb&site_no=01477050&legacy=&referred_module=sw&period=&begin_date=1961-12-27&end_date=2023-01-30"

drChester_header <- readLines(url,n=40) 

# Find where the header row is
hdrrow <- str_detect(drChester_header,"agency_cd") %>% which(.,TRUE)

drChester_header <- drChester_header %>% 
  .[hdrrow] %>% 
  strsplit(.,split="\t") %>% 
  unlist()

drChester <- read.table(url,sep='\t',skip=hdrrow+1, 
                        col.names = drChester_header) %>% 
  mutate(date=as.Date(datetime,format="%Y-%m-%d"))

# Subset the data to get daily mean spCond (µS/cm), wt (°C), and DO (mg/L)
drChester_ctdo <- drChester[,c(1,2,3,8,14,20)]
names(drChester_ctdo) <- c("agency_cd","site_no","datetime_chr","spCond.mean","wt.mean","do.mgL.mean")

# Convert the date and make DO a numeric variable 
drChester_ctdo <- drChester_ctdo %>% 
  mutate(date = as.Date(datetime_chr),
         do.mgL.mean=as.numeric(do.mgL.mean)) %>% 
  select(agency_cd,site_no,spCond.mean,wt.mean,do.mgL.mean,date)

# This is commented out so that the original data file is not overwritten. 
#save(drChester,drChester_ctdo,file=here("data","5. Chester_1961-2023.Rdata"))

```

## Plot the data for initial evaluation 

```{r tidy data, warning=FALSE}
# Load the Chester Time Series if desired.  
load(here("data","5. Chester_1961-2023.Rdata"))

# Plot the Chester Data for initial evaluation to ensure the data is as expected
tmp <- drChester_ctdo %>% 
  mutate(year=year(date)) %>% 
  select(date,year,wt.mean,spCond.mean,do.mgL.mean) 

# Calculate salinity and percent oxygen saturation
tmp <- tmp[complete.cases(tmp),]
tmp <- tmp %>% mutate(salinity.mean = ec2pss(ec=spCond.mean/1000,t=wt.mean,p = 0))

# Calculate percent oxygen saturation.
# If salinity is less than 1, use APHA method that does not use salinity. APHA
# method is the standard for fresh water application.
# If salinity is >=1, then use "Weiss" method, which considers salinity, and is 
# the standard formula in marine sciences (Soetert and Petzoldt 2020).
# Soetaert K, Petzoldt T (2020). _marelac: Tools for Aquatic Sciences_. R package
# version 2.1.10, <https://CRAN.R-project.org/package=marelac>.
tmp <- tmp %>% mutate(salinity.mean=ec2pss(ec=spCond.mean/1000,t=wt.mean),
      posat.mean=ifelse(salinity.mean<1,
      do.mgL.mean*100/gas_O2sat(t=wt.mean,method="APHA"),
      do.mgL.mean*100/gas_O2sat(S=salinity.mean,t=wt.mean,method="Weiss")
      ))

ggplot(filter(tmp,year>2002),aes(x=date,y=posat.mean,color=wt.mean))+
  geom_point()+
  scale_color_continuous(type="viridis",breaks=seq(5,25,by=5))+
  scale_x_date(date_breaks="2 years",date_labels="%m/%y")+
  labs(x="",y="Percent DO Saturation")+
  theme(legend.title=element_blank())+
  no_grid
```

## Plot POSAT for 20 years at Chester
Create a facet plot showing each year of DO data from 2002 to 2022. This section creates Figure A1-1 in the Technical Support Document.

```{r Plot 20 Years of Data}
## Plot 20 years of DO at Chester.

# recast the dates as a date in 2002 so that same scale can be used in all the
# facets.  Dotted lines show Jul 1-Oct 31 which is the simulation period for the
# fish model.
plt <- tmp %>% filter(between(year(date),2002,2022)) %>% 
  mutate(dy=as.Date(paste0("2002-",format(date,format="%m-%d")))) %>% 

ggplot(aes(x=dy,y=posat.mean,color=wt.mean))+
  geom_point()+
  #geom_line()+
  facet_wrap(~year(date),ncol=4)+
  scale_color_continuous(type="viridis",breaks=seq(5,25,by=5))+
  scale_x_date(
    limits=c(as.Date("2002-06-01"),as.Date("2002-11-30")),
    breaks=as.Date(c("2002-06-01","2002-07-01","2002-08-01",
        "2002-09-01","2002-10-01","2002-11-01")),
    labels=c("","7/1","","9/1","","11/1"))+
  labs(x="",y="Percent DO Saturation")+
  ggtitle("Chester, PA Dissolved Oxygen, 2002-2022")+
  geom_vline(xintercept=c(as.Date("2002-07-01"),as.Date("2002-10-31")),linetype=2)+
  no_grid+
  theme(legend.position="inside",legend.position.inside=c(0.6,0.05),legend.direction="horizontal")+
  guides(color=guide_colorbar(title="Water Temperature"))
plt

# Creates Figure A1-1 of the TSD.
 pdf(file=here("figures","5. 2002-2022 POSAT Time Series.pdf"),
     width=6,height=8,useDingbats = FALSE,paper="letter")
   plt
 dev.off()

```

## Generate Summer Complete Data
This code chunk uses linear interpolation to impute values for any missing data between July 1 and November 30.  "Complete" refers to the absence of any missing values from the daily time series, which is required for later use in the cohort model.

```{r Generate Summer Complete Data}

drc <- drChester_ctdo 

# Eliminate rows with missing values
drc <- drc[complete.cases(drc),] %>% 
  mutate(year=year(date))

# Make a summer time series for each year from 2002 to 2022
Year <- seq(2002,2022,by=1)
# Define data to be imputed
ts_input <- drc  
# Impute the data
drc_summer <- do.call(rbind,lapply(Year,imputeByYear))

# Don't overwrite the data file
#save(drc_summer,file=here("data","5. DR_Chester_Summer_Complete.Rdata"))

```

## Make a graph of long term DO concentration for CERF talk

```{r Make Graph for CERF, eval=FALSE, include=FALSE}

plt <- tmp %>% 
  ggplot(aes(x=date,y=posat.mean,color=wt.mean))+
  geom_point(size=3,shape=1,alpha=0.5)+
  scale_color_viridis(discrete=FALSE)+
  scale_y_continuous(breaks=seq(0,120,by=20))+
  scale_x_date(limits=c(as.Date("1965-01-01"),as.Date("2025-01-01")),
    breaks=as.Date(paste0(seq(1965,2025,by=10),"-01-01"),
                   format="%Y-%m-%d"),date_labels="%Y")+
  labs(x="",y="Percent Oxygen Saturation",color="Water\nTemperature\n(°C)")+
  theme_classic()+
  large_text+
  theme(legend.text=element_text(size=14),
        legend.title = element_text(size=14))
plt

png(file=here("figures","5. POSAT Time Series.png"),width=8,height=4,units="in",res=150)
  plt
dev.off()

```
