==> R/5.02 Visualize WQ at Chester and Penn's Landing.Rmd <==
---
title: "5.02 Visualize WQ Distribution at Chester and Penn's Landing.Rmd"
author: "J. Hagy"
date: "`r Sys.Date()`"
output:
  html_document:
    df_print: paged
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(marelac)
library(wql)
library(wesanderson)
library(viridis)
library(here)
library(patchwork)
library(qacBase)

no_grid <- theme(panel.grid=element_blank())

source(here("functions","Atlantic Sturgeon Growth Model Function v3 (2023-06-13).R"))

pal <- wes_palette("Zissou1",5,type="discrete")

```

## Load data for summers

```{r load data}

load(here("data","5. DR_Chester_Summer_Complete.Rdata"))
load(here("data","5.01 DR_Penn_Landing_Summer_Complete 2002-2022.Rdata"))

df <- rbind(
  drc_summer %>% mutate(site="Chester"),
  drpl_summer %>% mutate(site="Penn's Landing")
)

df <- df %>% 
  filter(complete.cases(.)) %>% 
  select(site,date,doy,year,wt.mean,salinity.mean,posat.mean) %>% 
  mutate(do.mgl=posat.mean*gas_O2sat(S=salinity.mean,t=wt.mean)/100) %>% 
  pivot_longer(cols=all_of(c("wt.mean","salinity.mean","posat.mean","do.mgl")),
               names_to = "param",
               values_to = "value") %>% 
  mutate(site=factor(site),param=factor(param))

# Create a date variable and re-order the parameter factor levels
df <- df %>% 
  mutate(doy.as.date = as.Date(paste0(2004,format(date,format="-%m-%d")))) %>% 
  mutate(param=factor(param,levels=c("salinity.mean","wt.mean","posat.mean","do.mgl"),
                      labels=c("Salinity","Water Temperature",
                               "Percent Oxygen Saturation","Dissolved Oxygen"))
         )

# Calculate the median by day of year
df.med <- df %>% group_by(site,param,doy.as.date) %>% 
  summarize(med = median(value)) %>% 
  ungroup()
  
```

## Make a graph of seasonal WQ distribution

```{r Graph Seasonal WQ Distribution, echo=TRUE}

graphClimatology <- function(myParam,ylabel,xInterval="1 month") {

df <- filter(df,param==myParam)
df.med <- filter(df.med,param==myParam)

plt <- ggplot()+
  geom_point(data=df,aes(x=doy.as.date,y=value),shape=1,alpha=0.2)+
  geom_line(data=df.med,aes(x=doy.as.date,y=med),color="red")+
  facet_wrap(~fct_rev(site))+
  scale_x_date(date_breaks=xInterval,date_labels = "%m/%d")+
  geom_vline(xintercept = c(as.Date("2004-07-01"),as.Date("2004-10-31")),linetype=2)+
  no_grid+
  labs(x="",y=ylabel)
return(plt)
}

plt.sal <- graphClimatology("Salinity","Salinity (ppt)")
plt.wt <- graphClimatology("Water Temperature","Water Temperature (°C)")
plt.posat <- graphClimatology("Percent Oxygen Saturation","Percent Oxygen Saturation")
plt.do <- graphClimatology("Dissolved Oxygen","Dissolved Oxygen (mg/L)")

# Plot an 8-panel figure
plt.sal / plt.wt / plt.posat / plt.do

# This creates figure 4 in the ORD Report, but is not used in the TSD
pdf(file=here("figures","5.02 Seasonal WQ Distribution.pdf"),
    useDingbats = FALSE,width=6,height=8,paper="letter")
  plt.sal / plt.wt / plt.posat / plt.do
dev.off()
```

# Summarize Water Quality statistics

```{r Summarize Statistics, echo=TRUE}

qstats(df,value,site, param, stats = c("min","median","mean","max","sd","n"))

```

# Make a contour plot showing effect of observed WQ on growth rate

```{r Contour Growth Rate, echo=TRUE}

dat <- expand.grid(posat=seq(30,120),wt=seq(10,31,by=0.5))
dat$growth <- with(dat,as_growth_model3(TEMP=wt,SAL=0.5,DO=posat,GR=50))

df <- rbind(
  drc_summer %>% mutate(site="Chester"),
  drpl_summer %>% mutate(site="Penn's Landing")) %>%
  filter(between(doy,184,304)) %>% 
  mutate(site=factor(site,levels=c("Penn's Landing","Chester")))

plt <- ggplot()+
  geom_contour_filled(data=dat,aes(x=posat,y=wt,z=growth))+
  geom_contour(data=dat,aes(x=posat,y=wt,z=growth),breaks=0,color=pal[5],linewidth=1)+
  geom_point(data=df[seq(1,nrow(df),by=10),],aes(x=posat.mean,y=wt.mean),size=2,shape=1,color="#000000")+
  theme(panel.grid=element_blank())+
  facet_wrap(~site,ncol=1)+
  scale_x_continuous(breaks=seq(30,120,by=10),expand=c(0,0))+
  scale_y_continuous(breaks=seq(10,30,by=5),expand=c(0,0))+
  labs(x="Percent Oxygen Saturation",y="Water Temperature (°C)")+
  theme(strip.text=element_text(size=10),
        axis.text=element_text(size=10),
        axis.title=element_text(size=12))+
  guides(fill=guide_legend(title="Growth Rate\n(per day)"))
plt

# This creates figure 5 in the TSD
pdf(file=here("figures","5.02 Filled contours with observed WQ.pdf"),
    height=5,width=5,paper="letter")
  plt
dev.off()

```

# Make plots similar to above for presentation purposes
This graph is very similar to the graph created above, but it is formatted to look better in a powerpoint presentation

```{r Growth Contour for CERF, eval=FALSE, include=FALSE}

dat <- expand.grid(posat=seq(30,120),wt=seq(0,31,by=0.5),site=c("Chester","Penn's Landing"))
dat$growth <- with(dat,as_growth_model3(TEMP=wt,SAL=0.5,DO=posat,GR=50))

df <- rbind(
  drc_summer %>% mutate(site="Chester"),
  drpl_summer %>% mutate(site="Penn's Landing")) %>%
  filter(between(doy,184,304)) %>% 
  mutate(site=factor(site,levels=c("Penn's Landing","Chester")))

plt.noPoints <- ggplot() +
  geom_contour_filled(data=dat,aes(x=posat,y=wt,z=growth))+
  geom_contour(data=dat,aes(x=posat,y=wt,z=growth),breaks=0,color=pal[5],linewidth=1)+
  theme(panel.grid=element_blank())+
  facet_wrap(~site,ncol=1)+
  scale_x_continuous(breaks=seq(30,120,by=20),expand=c(0,0))+
  scale_y_continuous(breaks=seq(5,30,by=5),expand=c(0,0))+
  labs(x="Percent Oxygen Saturation",y="Water Temperature (°C)")+
  theme(strip.text=element_text(size=10),
        axis.text=element_text(size=10),
        axis.title=element_text(size=12))+
  guides(fill=guide_legend(title="Growth Rate\n(per day)"))+
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14),
        strip.text=element_text(size=14))
plt.noPoints

pdf(file=here("figures","5.02 Filled growth contours for CERF talk.pdf"),height=5,width=6,paper="letter")
  plt.noPoints
dev.off()

plt <- ggplot()+
  geom_contour_filled(data=dat,aes(x=posat,y=wt,z=growth))+
  geom_contour(data=dat,aes(x=posat,y=wt,z=growth),breaks=0,color=pal[5],linewidth=1)+
  geom_point(data=df[seq(1,nrow(df),by=10),],aes(x=posat.mean,y=wt.mean),size=2,shape=1,color="#000000")+
  theme(panel.grid=element_blank())+
  facet_wrap(~site,ncol=1)+
  scale_x_continuous(breaks=seq(30,110,by=20),expand=c(0,0))+
  scale_y_continuous(breaks=seq(5,30,by=5),expand=c(0,0))+
  labs(x="Percent Oxygen Saturation",y="Water Temperature (°C)")+
  theme(strip.text=element_text(size=10),
        axis.text=element_text(size=10),
        axis.title=element_text(size=12))+
  guides(fill=guide_legend(title="Growth Rate\n(per day)"))+
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14),
        strip.text=element_text(size=14))
plt

pdf(file=here("figures","5.02 Filled contours with observed WQ for CERF talk.pdf"),
    height=5,width=6,paper="letter")
  plt
dev.off()


```

# Plot full year WQ distribution for Technical Support Document

```{r}

load(here("data","5.01 Delaware River at Penn Landing.Rdata"))
load(here("data","5. Chester_1961-2023.Rdata"))

# Combine data frames and eliminate rows with missing values
df <- rbind(
  drChester_ctdo %>% 
    select(-agency_cd,-site_no) %>% 
    mutate(site="Chester"),
  drPenn_ctdo %>% 
    select(-agency_cd,-site_no) %>% 
    mutate(site="Penn Landing")
) %>% filter(complete.cases(.))

# Calculate salinity and POSAT
df <- df %>% 
  mutate(salinity.mean = ec2pss(ec=spCond.mean/1000,t=wt.mean,p = 0),
         posat.mean=do.mgL.mean*100/gas_O2sat(S=salinity.mean,t=wt.mean))

# Calculate time variables
df <- df %>% 
  mutate(doy=yday(date),
         cyear=year(date)+doy/366)

# Only keep data for 2002 and later
df <- df %>% 
  filter(year(date)>2001)

# Transpose data
df <- df %>% 
  mutate(year=year(date)) %>% 
  mutate(doy.as.date = as.Date(paste0(2004,format(date,format="-%m-%d")))) %>% 
  select(site,date,doy.as.date,year,wt.mean,salinity.mean,posat.mean) %>% 
  mutate(do.mgl=posat.mean*gas_O2sat(S=salinity.mean,t=wt.mean)/100) %>% 
  pivot_longer(cols=all_of(c("wt.mean","salinity.mean","posat.mean","do.mgl")),
               names_to = "param",
               values_to = "value") %>% 
  mutate(site=factor(site),param=factor(param))

# Create a date variable and re-order the parameter factor levels
df <- df %>% 
  mutate(param=factor(param,levels=c("salinity.mean","wt.mean","posat.mean","do.mgl"),
                      labels=c("Salinity","Water Temperature",
                               "Percent Oxygen Saturation","Dissolved Oxygen"))
  )

# Calculate the median by day of year
df.med <- df %>% group_by(site,param,doy.as.date) %>% 
  summarize(med = median(value)) 

# Make climatology graphs
plt.sal <- graphClimatology("Salinity","Salinity (ppt)",xInterval="2 months")
plt.wt <- graphClimatology("Water Temperature","Water Temperature (°C)",xInterval="2 months")
plt.posat <- graphClimatology("Percent Oxygen Saturation","Percent Oxygen Saturation",xInterval="2 months")
plt.do <- graphClimatology("Dissolved Oxygen","Dissolved Oxygen (mg/L)",xInterval="2 months")

# Plot an 8-panel figure
plt.sal / plt.wt / plt.posat / plt.do

# This creates Figure 2 in the TSD
pdf(file=here("figures","5.02 Full Year WQ Distribution.pdf"),
      useDingbats = FALSE,width=6,height=8,paper="letter")
  plt.sal / plt.wt / plt.posat / plt.do
dev.off()

```

# Calculate distribution statistics by Season

```{r}

df <- df %>% mutate(
  season=cut(doy.as.date,
  breaks=c(as.Date("2003-12-31"),as.Date("2004-03-01"),
    as.Date("2004-07-01"),as.Date("2004-11-01"),
    as.Date("2005-01-02")),
  labels=c("Overwinter","Spawning","Growth","Overwinter"))
)

# Calculate seasonal descriptive statistics
df.summary <- df %>% 
  filter(!is.na(value)) %>% 
  filter(date >= as.Date("2011-11-01")) %>% 
  group_by(site,param,season) %>% 
  summarize(min=min(value),
            p.10=quantile(value,probs=0.10),
            p.25=quantile(value,probs=0.25),
            med=median(value),mean=mean(value),
            p.75=quantile(value,probs=0.75),
            p.90=quantile(value,probs=0.90),
            max=max(value),
            n=n())
df.summary

# Calculate seasonal stats on the median seasonal climatology
df.med <- df.med %>% 

stats_seasonal_climatology <- df.med %>% 
  mutate(season=cut(doy.as.date,
      breaks=c(as.Date("2003-12-31"),as.Date("2004-03-01"),
          as.Date("2004-07-01"),as.Date("2004-11-01"),
          as.Date("2005-01-02")),
      labels=c("Overwinter","Spawning","Growth","Overwinter"))
  ) %>% 
group_by(site,season,param) %>% 
  summarize(
  p.90 = quantile(med,probs=0.9),
  mean = mean(med)
  )
stats_seasonal_climatology

write.csv(df.summary,file=here("outputs","5.02 seasonal WQ stats.csv"))
write.csv(stats_seasonal_climatology,file=here("outputs","5.02 seasonal stats on median climatology.csv"))

# Calculate seasonal descriptive statistics for just 2012-2022, 
df.summary <- df %>% 
  filter(year>=2012) %>% 
  filter(!is.na(value)) %>% 
  group_by(param,site,season) %>% 
  summarize(min=min(value),p.25=quantile(value,probs=0.25),
            med=median(value),mean=mean(value),
            p.75=quantile(value,probs=0.75),max=max(value))


write.csv(df.summary,file=here("outputs","5.02 seasonal WQ stats.2012-2022.csv"))

```
