==> R/5.04 Simulate Cohort for Penns Landing.Rmd <==
---
title: "5.04 Simulate Cohort for 2002-2022 at Penn's Landing"
author: "J. Hagy"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(wesanderson)
library(viridis)
library(here)

# Load functions implementing bioenergetics model

source(here("functions","Atlantic Sturgeon Growth Model Function v3 (2023-06-13).R"))
source(here("functions","get_mortality.R"))

basic_theme <- theme_classic() +
  theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14))
large_text <- theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14))
nogrid <- theme(panel.grid=element_blank())

pal <- wes_palette("Zissou1",5,type="discrete")

```

# Cohort Simulation at Penn's Landing

This code follows the cohort modeling concept of Niklitschek and Secor (2009) to simulate growth and survival of a juvenile cohort experiencing water quality as observed at Penn's Landing, PA

```{r Simulate Cohort}

# Load the summer water quality data for Penn's Landing
load(here("data","5.01 DR_Penn_Landing_Summer_Complete 2002-2022.Rdata"))

# Prepare water quality matrices for fish model for use in 
# simulations
posat.mean <- drpl_summer %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,posat.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = posat.mean) %>% 
  as.matrix()

salinity.mean <- drpl_summer %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,salinity.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = salinity.mean) %>% 
  as.matrix()

wt.mean <- drpl_summer %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,wt.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = wt.mean) %>% 
  as.matrix()

numberOfYears <- nrow(posat.mean)

# Subset the posat,sal,wt matrices to start on July 1 following 
posat <- posat.mean[,32:ncol(posat.mean)]
wt <- wt.mean[,32:ncol(wt.mean)]
salinity <- salinity.mean[,32:ncol(salinity.mean)]

# Simulate cohort
initial.weight <- 27  # 27 grams
weight <- array(0,dim=c(numberOfYears,154))
abundance <- array(0,dim=c(numberOfYears,154))
growth <- array(0,dim=c(numberOfYears,154))
mortality <- array(0,dim=c(numberOfYears,154))
weight[,1] <- initial.weight
abundance[,1] <- 10000

for (t in 1:153) {
  G <- as_growth_model3(TEMP=wt[,t],DO=posat[,t],SAL=salinity[,t],GR=weight[,t])
  m.size <- 0  # set natural mortality to zero.
  Z.do <- get.m.hypoxia(POSAT=posat[,t],WT=wt[,t])
  weight[,t+1] <- weight[,t]*exp(G)
  abundance[,t+1] <- abundance[,t]*exp(-m.size-Z.do)
  growth[,t] <- G
  mortality[,t] <- Z.do
  # if weight falls below initial weight, reset to initial weight. This prevents weight
  # from falling below minimum size for growth model to return a value.
  weight[weight[,t+1]<initial.weight,t+1] <- initial.weight
}

```

## Combine results
This code processes the simulation model results from the previous code chunk so that it can be graphed and used in later analysis.

```{r Combine Simulation Results}

yrs <- unique(drpl_summer$year)

transformArray <- function(inArray,vname) {
  doy <- seq.Date(from=as.Date("2002-07-01"),to=as.Date("2002-12-01"),by="1 day") %>% format(.,format="-%m-%d")
  rownames(inArray) <- paste0("Y",yrs)
  sim <- t(inArray) %>% 
    as.data.frame() %>% 
    cbind(doy,.) %>% 
    pivot_longer(cols = starts_with("Y"),names_prefix = "Y",names_to="year",values_to = vname) 
  sim <- sim[order(sim$year,sim$doy),]
}

sim.wt <- transformArray(weight,"weight")
sim.abund <- transformArray(abundance,"abundance")
sim.G <- transformArray(growth,"growth")
sim.Z <- transformArray(mortality,"mortality")

sim.PennsLanding.Zmin <- cbind(sim.abund,weight=sim.wt$weight,growth=sim.G$growth,mortality=sim.Z$mortality) %>% 
  mutate(date = as.Date(paste0(year,doy))) %>% 
  select(-doy) %>% 
  mutate(year=as.numeric(year))

# Save the cohort simulation data
#save(sim.PennsLanding.Zmin,
#     file=here("data","5.04 Penns Landing Cohort Simulation 2002-2022.Rdata"))

```

## Make a graph of fish size and compare with observed sizes

```{r Compare Simulated and Observed Fish Size}

#load(paste0(dataPath,"5.04 Penns Landing Cohort Simulation 2002-2022.Rdata"))

# Load data set from Ian Park with Sturgeon sizes
load(here("data","3.0 size.on.date.Rdata"))

# Graph fish weight
size.on.date <- size.on.date %>% 
  mutate(date.2019=as.Date(yday(date),origin=as.Date("2018-12-31")))

# Create a variable indicating if year had fish captures in Ian
# Park's data or not.
df <- sim.PennsLanding.Zmin %>% 
  mutate(year.with.capture = ifelse(year %in% unique(size.on.date$year.capture),"With Capture","No Captures")) %>% mutate(year.with.capture=factor(year.with.capture,levels=c("With Capture","No Captures")))

plt <- ggplot()+
  geom_line(data=df,aes(x=as.Date(paste0("2019-",format(date,"%m-%d"))),
      y=weight,group=factor(year),lwd=year.with.capture))+
  geom_point(data=filter(size.on.date,yday(date)>150),
             aes(x=doy.as.date,y=weight.g,color=source),size=3,alpha=0.5)+
  scale_x_date(date_breaks="1 months",date_labels="%b")+
  scale_linewidth_manual(values = c(1,0.3),guide="none")+
  scale_y_continuous(breaks=seq(0,250,50))+
  scale_color_manual(values=pal[c(1,5)])+
  coord_cartesian(ylim=c(0,225))+
  labs(x="",y="Fish Weight (g)")+
  basic_theme+
  theme(legend.position=c(0.15,0.6))+
  guides(color=guide_legend("Source"))
plt

# This figure is not included in the TSD.
pdf(file=here("figures","5.04 Predicted fish weight at PennsLanding.pdf"),
     width=4.5,height=4,paper="letter")
  plt
dev.off()

```

## Graph Growth Rate (Gmax) and Mortality Rate (Zmin)

```{r Graph Growth and Mortality Rate}

# Load previously saved data.
#load(paste0(dataPath,"5.03 Penns Landing Cohort Simulation 2002-2022.Rdata"))

## Remove data from 2010 (too much missing data, per DRBC)
sim.PennsLanding.Zmin <- sim.PennsLanding.Zmin %>% filter(year != "2010" )

# Graph growth
sim.PennsLanding.Zmin %>% 
  mutate(doy.as.date=as.Date(paste("2019-",format(date,"%m-%d")))) %>%
  select(doy.as.date,year,growth,mortality) %>% 
  pivot_longer(cols=all_of(c("growth","mortality")),names_to="param",values_to = "rate") %>% 
  mutate(param=factor(param,levels=c("growth","mortality"),labels=c("Growth","Mortality"))) %>% 
  filter(year != 2010) %>% 
ggplot(aes(x=doy.as.date,y=rate,color=factor(year)))+
  facet_wrap(~param,scales="free_y",ncol=1)+
  geom_line(linewidth=0.7)+
  geom_hline(yintercept=0,linetype=2)+
  guides(color=guide_legend(title="Year"))+
  scale_x_date(date_breaks = "1 month",date_labels = "%m/%d")+
  labs(x="",y="Instantaneous Rate (/d)")+
  nogrid+large_text

# graph Gmax minus Zmin (phi)
sim.PennsLanding.Zmin %>% 
  mutate(doy.as.date=as.Date(paste("2019-",format(date,"%m-%d"))),
         phi=growth-mortality) %>%
  filter(year != 2010) %>% 
ggplot(aes(x=doy.as.date,y=phi))+
  facet_wrap(~year,ncol=4)+
  geom_line(linewidth=0.7)+
  geom_hline(yintercept=0,linetype=2)+
  guides(color=guide_legend(title="Year"))+
  scale_x_date(date_breaks = "3 month",date_labels = "%m/%d")+
  labs(x="",y="Production Potential (/d)")+
  nogrid+
  ggtitle("Production Potential at Penn's Landing, PA, 2002-2022")

```

## Graph Mean Production Potential for 2002-2022

```{r}

df <- sim.PennsLanding.Zmin %>% 
  mutate(phi=growth-mortality) %>% 
  filter(between(month(date),7,10)) %>% 
  group_by(year) %>% 
  summarize(mean.phi=mean(phi),mean.growth=mean(growth),mean.Zmin=mean(mortality))


ggplot(df,aes(x=year,y=mean.phi))+
  geom_col(fill=pal[1])+
  geom_hline(yintercept=0,linetype=2)+
  scale_y_continuous(limits=c(-0.04,0.01),breaks=seq(-0.04,0.01,by=0.01))+
  scale_x_continuous(breaks=seq(2002,2022,by=2))+
  basic_theme+
  labs(y="Potential Production (/d)",x="")+
  ggtitle("Production Potential at Penn's Landing, PA",
          subtitle="July 1 - November 1")

```
