==> R/6.01 Simulate Cohorts using Restored DO.Rmd <==
---
title: "6.01 Simulate Cohorts Using Restored DO.Rmd"
author: "J. Hagy"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(marelac)
library(wesanderson)
library(viridis)
library(here)

# Load functions implementing bioenergetics model
source(here("functions","Atlantic Sturgeon Growth Model Function v3 (2023-06-13).R"))
source(here("functions","get_mortality.R"))

large_text <- theme(axis.text=element_text(size=14),
        axis.title=element_text(size=14),
        strip.text=element_text(size=14))
no_grid <- theme(panel.grid=element_blank())

pal <- wes_palette("Zissou1",5,type="discrete")

```

# Overview
Program 6.0 modeled the relationship between simulated DO from DRBC's HADO scenario and observed DO to predict restored DO for every year from 2002-2022. In this program, we simulate the daily instantaneous potential growth rate and daily instantaneous minimum mortality rate using the restored DO estimates.  We then calculate Habitat Suitability Index (HSI), which is defined as the July 1 to October 31 average of potential growth rate minus  minimum mortality rates.

## Load Restored DO estimates
```{r Load Restored DO, echo=TRUE}

load(here("data","6.0 Estimates of Restored DO.Rdata"))

```

## Simulate HSI using Restored DO

```{r Simulate HSI}

#Make a function to implement the simulation 
sim <- function(mySite) {

df <- restoredWQ %>%
  select(Site,year,date,posat.mean,sal.mean,wt.mean) %>% 
  filter(between(month(date),7,10)) %>% 
  filter(Site==mySite) %>% 
  filter(year != 2010)

posat <- df %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,posat.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = posat.mean) %>% 
  select(-year) %>% 
  as.matrix()

salinity <- df %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,sal.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = sal.mean) %>% 
  select(-year) %>% 
  as.matrix()

wt <- df %>%
  mutate(doy = format(date,"%m%d")) %>% 
  select(year,doy,wt.mean) %>% 
  pivot_wider(names_from=doy,names_prefix="day",values_from = wt.mean) %>% 
  select(-year) %>% 
  as.matrix()

numberOfYears <- nrow(posat)

# Simulate cohort
initial.weight <- 27
weight <- array(0,dim=c(numberOfYears,124))
abundance <- array(0,dim=c(numberOfYears,124))
growth <- array(0,dim=c(numberOfYears,124))
mortality <- array(0,dim=c(numberOfYears,124))
weight[,1] <- initial.weight
abundance[,1] <- 10000

for (t in 1:123) {
  G <- as_growth_model3(TEMP=wt[,t],DO=posat[,t],SAL=salinity[,t],GR=weight[,t])
  m.size <- 0  # set natural mortality to zero.
  Z.do <- get.m.hypoxia(POSAT=posat[,t],WT=wt[,t])
  weight[,t+1] <- weight[,t]*exp(G)
  abundance[,t+1] <- abundance[,t]*exp(-m.size-Z.do)
  growth[,t] <- G
  mortality[,t] <- Z.do
  # if weight falls below initial weight, then reset to initial weight
  weight[weight[,t+1]<initial.weight,t+1] <- initial.weight  
}

yrs <- unique(df$year)

transformArray <- function(inArray,vname) {
  doy <- seq.Date(from=as.Date("2002-07-01"),to=as.Date("2002-11-01"),by="1 day") %>% format(.,format="-%m-%d")
  rownames(inArray) <- paste0("Y",yrs)
  sim <- t(inArray) %>% 
    as.data.frame() %>% 
    cbind(doy,.) %>% 
    pivot_longer(cols = starts_with("Y"),names_prefix = "Y",names_to="year",values_to = vname) 
  sim <- sim[order(sim$year,sim$doy),]
}

sim.wt <- transformArray(weight,"weight")
sim.abund <- transformArray(abundance,"abundance")
sim.G <- transformArray(growth,"growth")
sim.Z <- transformArray(mortality,"mortality")

sim.combined <- cbind(sim.abund,weight=sim.wt$weight,growth=sim.G$growth,mortality=sim.Z$mortality) %>% 
    mutate(year=as.numeric(year),
           date=as.Date(paste0(year,doy))
    )

summary <- sim.combined  %>% 
  mutate(phi=growth-mortality) %>% 
  mutate(year=as.numeric(year)) %>% 
  filter(between(month(date),7,10)) %>% 
  group_by(year) %>% 
  summarize(mean.phi=mean(phi),mean.growth=mean(growth),mean.Zmin=mean(mortality)) %>% 
  mutate(Site=mySite)

return(summary)
}

# Run the simulations for all the zones and locations, and bind them together in a data frame.
results <- rbind(
  sim("Penn's Landing"),
  sim("Chester")
  )

restoredHSI <- results
rm(results)
save(restoredWQ,restoredHSI,
     file=here("data","6.01 Simulated HSI using Restored DO.Rdata"))
```

# Plot HSI Results

```{r Plot HSI Results}

# Plot histograms of the restored DO values.  Note that values below 70% are generally absent at Chester, with most values between 75% and 90%.
restoredWQ %>% 
  ggplot(aes(x=posat.mean))+
  geom_histogram()+
  facet_grid(~Site)

# Don't include 2010 data because there is too much missing data
restoredHSI <- filter(restoredHSI,year != 2010)

ggplot(restoredHSI,aes(x=year,y=mean.phi,fill=Site))+
  geom_col(position="dodge")+
  geom_hline(yintercept=0)+
  no_grid+
  ggtitle("Calculated Habitat Suitability Index using Restored DO") 

```
