==> R/7.04 Compare Seasonal Minimum and 10th Percentile.Rmd <==
---
title: "7.0 Compare Seasonal Minimum and 10th Percentile"
author: "J Hagy"
date: "2024-09-04"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(here)
library(wesanderson)

pal <- wes_palette("Zissou1",5,"discrete")

# Load the existing percentiles
load(here("data","6.02 DO and POSAT percentiles.Rdata"))
# Percentile file does not include the seasonal minimum

# Load the summer complete data
load(here("data","5.01 DR_Penn_Landing_Summer_Complete 2002-2022.Rdata"))
load(here("data","5. DR_Chester_Summer_Complete.Rdata"))

# Load the restored DO data
load(here("data","6.0 Estimates of Restored DO.Rdata"))

```

This program evaluates the relationship between values of the 10th percentile and values of the seasonal minimum POSAT, generating Figure 10 in the TSD for the final rule.


## Calculate the annual minimums
Add the minimum to the percentiles, using percentile=0.  Do this for both the observed DO and restored DO.

```{r}

df <- rbind(drc_summer %>% mutate(Site="Chester"),
            drpl_summer %>% mutate(Site="Penn's Landing")) %>% 
  filter(between(month(date),7,10))

mins <- df %>% group_by(Site,year) %>% 
  summarize(posat=min(posat.mean),
            do=min(do.mgL.mean)) %>% 
  ungroup() %>% 
  mutate(type="Observed",percentile=0) %>% 
  select(type,Site,year,do,posat,percentile)

mins.res <- restoredWQ %>% 
  filter(between(month(date),7,10)) %>% 
  group_by(Site,year) %>%
  summarize(posat=min(posat.mean),
            do=min(do.mean)) %>% 
  ungroup() %>% 
  mutate(type="Restored",percentile=0) %>% 
  select(type,Site,year,do,posat,percentile)

df <- rbind(percentiles,mins,mins.res)

```

## Graph minimum vs. 10th percentile
Include predicted minimum when then 10th percentile POSAT is 66%

```{r}

df.pivot <- df %>% 
  mutate(percentile=percentile*100) %>% 
  select(type,Site,year,posat,percentile) %>% 
  pivot_wider(names_from = percentile,names_prefix = "P.",
              values_from = posat)

fit <- lm(P.0 ~ P.10,data=df.pivot)
summary(fit)
tmp <- predict(fit,data.frame(P.10=66),se.fit=TRUE)
tmp
minCI <- tmp$fit+qnorm(c(0.025,0.975),0,tmp$se.fit)

plt <- ggplot() +
  geom_point(data=df.pivot,aes(x=P.10,y=P.0,color=Site,shape=type),size=4)+
  geom_smooth(data=df.pivot,aes(x=P.10,y=P.0),method="lm")+
  geom_vline(xintercept = 66,linetype=2)+
  geom_hline(yintercept=tmp$fit,linetype=2)+
  theme(panel.grid=element_blank(),
        legend.position="top")+
  labs(x="10th Percentile Oxygen Saturation",y="Minimum Percent Oxygen Saturation")
plt

```

# Expected minimum DO values
If the minimum 10th percentile POSAT of 66% is to be attained in the Juvenile Development season in every year, then the 10th percentile will likely be higher than 66% in most or all years. 

Given an estimate of the standard deviation of the percentiles, we can estimate the mean percentile assuming that the percentiles are normally distributed.  Using data from Chester, the percentiles are approximately normally distributed.

```{r}

tmp <- percentiles %>% 
  filter(Site=="Penn's Landing") %>% 
  filter(type=="Observed")
qqnorm(tmp$posat)
qqline(tmp$posat)

tmp <- df %>% 
  filter(percentile==0.1) %>% 
  group_by(Site,type) %>% 
  summarize(sd=sd(posat))
tmp

```

What is the mean POSAT 10th percentile if 5% of 10th percentiles will be less than 66%?

```{r}

# t-statistic with 5% of distribution below.
qt(0.05,20)

tmp <- tmp %>% 
  mutate(meanP10=66-qt(0.05,20)*sd)
tmp

```

If the mean P.10 of POSAT is as calculated, what is the mean value for the minimum POSAT?  

```{r}

meanP0 <- predict(fit,data.frame(P.10=tmp$meanP10))

tmp <- tmp %>%
  ungroup() %>% 
  mutate(meanP.0=meanP0)
tmp

```

## Make a graph for TSD for the final rule

```{r}

plt <- ggplot() +
  geom_point(data=df.pivot,aes(x=P.10,y=P.0,color=Site,shape=type),size=4)+
  geom_smooth(data=df.pivot,aes(x=P.10,y=P.0),method="lm",fill=pal[2],alpha=0.2,color=pal[1])+
  # add vline for criterion
  geom_vline(xintercept = 66,linetype=2)+
  # draw confidence limit for prediction of minimum at P.10=66%
  geom_rect(aes(ymin=minCI[1],ymax=minCI[2],xmin=43,xmax=83),fill=pal[2],alpha=0.2)+
  geom_hline(yintercept=tmp$fit,linetype=2)+
  # draw 
  geom_rect(aes(ymin=37,ymax=82,xmin=71,xmax=73),fill=pal[3],alpha=0.2)+
  geom_rect(aes(ymin=67,ymax=69,xmin=43,xmax=83),fill=pal[3],alpha=0.2)+
  
  theme(panel.grid=element_blank(),
        legend.position="top")+
  labs(x="10th Percentile Oxygen Saturation",y="Minimum Percent Oxygen Saturation",
       shape="Scenario")+
  scale_x_continuous(limits=c(43,83),expand=c(0,0))+
  scale_y_continuous(limits=c(37,82),expand=c(0,0))+
  scale_color_manual(values=pal[c(1,5)])

plt2 <- plt+  
  annotate("text",x=44,y=68.2,hjust=0,label="Expected Mean")+
  annotate("text",x=71.8,y=38,hjust=0,angle=90,label="Expected Mean")+
  annotate("text",x=66.8,y=38,hjust=0,angle=90,label="66% Criterion")+
  annotate("text",x=44,y=62.8,hjust=0,label="Expected Minimum")+
  theme_classic()+
  theme(legend.position="inside",legend.position.inside=c(0.9,0.3))
plt2

pdf(file=here("figures","7.04 Expected Minimums Graph.pdf"),width=6,height=4,paper="letter")
  plt2
dev.off()

# Figure XX.  The relationship between the 10th percentile and minimum percent
# oxygen saturation during the Juvenile Development season at Chester and Penn’s
# Landing for observed values (2002-2022) and the restored scenario. A colored
# band (blue) shows the expected minimum percent oxygen saturation (with 95%
# confidence interval) when the 10th percentile is 66%.  Yellow bands show the
# expected mean 10th percentile and minimum value if the 66% criterion value is
# attained 95% of the time and variability is as expected in the restored DO
# scenario.
```

## Make graph like Fig 10 in TSD for the final rule, but in mg/L

```{r echo=TRUE}

df.pivot <- df %>% 
  mutate(percentile=percentile*100) %>% 
  select(type,Site,year,do,percentile) %>% 
  pivot_wider(names_from = percentile,names_prefix = "P.",
              values_from = do)

 fit <- lm(P.0 ~ P.10,data=df.pivot)
 summary(fit)
 tmp <- predict(fit,data.frame(P.10=5.4),se.fit=TRUE)
 tmp
 minCI <- tmp$fit+qnorm(c(0.025,0.975),0,tmp$se.fit)


plt <- ggplot() +
  geom_point(data=df.pivot,aes(x=P.10,y=P.0,color=Site,shape=type),size=4)+
  geom_smooth(data=df.pivot,aes(x=P.10,y=P.0),method="lm",fill=pal[2],alpha=0.2,color=pal[1])+
  # add vline for criterion
  geom_vline(xintercept = 5.4,linetype=2)+
  # draw confidence limit for prediction of minimum at P.10=66%
  geom_rect(aes(ymin=minCI[1],ymax=minCI[2],xmin=3.5,xmax=7),fill=pal[2],alpha=0.2)+
  geom_hline(yintercept=tmp$fit,linetype=2)+
  theme(panel.grid=element_blank(),
        legend.position="top")+
  labs(x="10th Percentile Dissolved Oxygen (mg/L)",y="Minimum Dissolved Oxygen (mg/L)",
       shape="Scenario")+
  scale_x_continuous(limits=c(3.5,7),expand=c(0,0))+
  scale_y_continuous(limits=c(3,6.5),expand=c(0,0))+
  scale_color_manual(values=pal[c(1,5)])+
  theme_classic()+
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14),
        legend.title = element_text(size=12))+
  theme(legend.position="inside",legend.position.inside=c(0.8,0.3))+
  annotate("text",x=5.5,y=3.5,hjust=0,angle=90,label="5.4 mg/L")+
  annotate("text",x=3.6,y=5.1,hjust=0,label="Expected Minimum")
plt
  
pdf(file=here("figures","7.04 10th Percentile vs Minimum DO mgL.pdf"),width=6,height=4,paper="letter")
  plt
dev.off()
