==> R/7.08 Evaluate NO23 in HADO.Rmd <==
---
title: "Evaluate NO3 in HADO"
author: "J Hagy"
date: "2025-08-07"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(here)
library(readxl)
library(wesanderson)
library(stats)
library(patchwork)

pal <- wes_palette("Zissou1",5,"discrete")
```

## Evaluate NO3 in HADO
This program provides The program evaluates the concentration of NO3 and NH4 in the HADO scenario to determine if it is possible that discharging NO3 from POTWs instead of ammonia nitrogen could result in NO3 exceeding the drinking water standard of 10 mg/L.

The data processing steps below read data from DRBC's EFDC/WASP Model (actual flows, kvg). Selecting results from zone 2 and 3 reduced the data storage requirements to a size that can be uploaded to GitHub. The resulting data file is 50 MB, which is much smaller than the original model output.

```{r}
# Read in the array
# dat <- read.csv(here("../ascohortmodel/raw/HADO_Actual_kavg","HADO_Actual_2019_NO3O2_kavg.csv"))
# dat.NH34 <- read.csv(here("../ascohortmodel/raw/HADO_Actual_kavg","HADO_Actual_2019_NH34_kavg.csv"))
# 
# # J index from 88 to 167 corresponds to the FMA
# 
# # Transpose the data
# dat.fma <- dat %>% 
#   pivot_longer(.,cols=starts_with("X"),names_to = "date_chr",values_to = "NO23") %>% 
#   mutate(datetime = as.POSIXct(date_chr,format="X%m.%d.%Y.%H.%M",tzone="America/New_York"))
# 
# # Transpose for Ammonium N
# dat.fma.NH34 <- dat.NH34 %>% 
#   pivot_longer(.,cols=starts_with("X"),names_to = "date_chr",values_to = "NH34") %>% 
#   mutate(datetime = as.POSIXct(date_chr,format="X%m.%d.%Y.%H.%M",tzone="America/New_York"))
# 
# # Combine the NO23 and NH34 data
# dat.fma <- cbind(dat.fma,dat.fma.NH34$NH34)
# dat.fma <- dat.fma %>% 
#   rename(NH34 = `dat.fma.NH34$NH34`)
# 
# # Filter for the FMA
# dat.fma$J <- str_split_i(dat.fma$IJ,"_",2) %>% as.numeric()
# summary(dat.fma$J)
# dat.fma <- filter(dat.fma,J>143)
# dat.fma$zone <- cut(dat.fma$J,breaks=c(142,177.5,264),labels=c(2,3))

# Reduce data to daily averages
# dat.fma <- dat.fma %>% 
#   select(zone,datetime,J,NH34,NO23) %>% 
#   mutate(date = as.Date(format(datetime,format="%Y-%m-%d"))) %>% 
#   group_by(zone,J,date) %>% 
#   summarize(NH34 = mean(NH34),
#             NO23 = mean(NO23))

# Save the data to a file.
#save(dat.fma,file=here("data","7.08 Nitrogen from HADO.Rdata"))

```

## Load the NH4 and NO23 model outputs and evaluate.

Determine the 95th percentile of each variable

```{r}

load(here("data","7.08 Nitrogen from HADO.Rdata"))

# By Zone
dat.fma %>% 
  group_by(zone) %>% 
  summarize(
    NO23.95 = quantile(NO23,probs=0.95,na.rm=TRUE),
    NO23.max = max(NO23,na.rm=TRUE)
  )

# Zones 2 and 3 together
dat.fma %>% 
  summarize(
    NO23.95 = quantile(NO23,probs=0.95,na.rm=TRUE),
    NO23.max = max(NO23,na.rm=TRUE)
  )

# Compute Dissolved Inorganic Nitrogen (DIN)
dat.fma$din <- dat.fma$NO23 + dat.fma$NH34

```

If the total of NH4 and NO23 is << 10 mg/L, it is unlikely that treatment of discharges to increase DO concentrations could result in NO3 > 10 mg/L.

```{r}

options(scipen = 999)
plt1 <- ggplot(dat.fma,aes(x=NO23))+
  geom_histogram(fill=pal[1],bins=50)+
  geom_vline(xintercept=10,linetype=2)+
  scale_x_log10(limits=c(0.001,100),breaks=c(0.01,0.1,1,10,100))+
  annotation_logticks(sides="b")+
  labs(x="Nitrate+Nitrite (mg/L)",y="Frequency")+
  theme_classic()+
  ggtitle("Inorganic Nitrogen in HADO Scenario (Zones 3, 4, upper-5")

plt2 <- ggplot(dat.fma,aes(x=NH34))+
  geom_histogram(fill=pal[1],bins=50)+
  geom_vline(xintercept=10,linetype=2)+
  scale_x_log10(limits=c(0.001,100),breaks=c(0.01,0.1,1,10,100))+
  annotation_logticks(sides="b")+
  labs(x="Ammonia Nitrogen (mg/L)",y="Frequency")+
  theme_classic()

plt3 <- ggplot(dat.fma,aes(x=din))+
  geom_histogram(fill=pal[1],bins=50)+
  geom_vline(xintercept=10,linetype=2)+
  scale_x_log10(limits=c(0.001,100),breaks=c(0.01,0.1,1,10,100))+
  annotation_logticks(sides="b")+
  labs(x="Dissolved Inorganic Nitrogen (mg/L)",y="Frequency")+
  theme_classic()

plt1 / plt2 / plt3
