==> R/7.09 Evaluate scenario where POSAT is less than 66 for 24 days.Rmd <==
---
title: "Evaluate 24 days of non attainment"
author: "J Hagy"
date: "2024-09-16"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(quantreg)
library(here)
library(betareg)

# Load the existing percentiles
load(here("data","6.02 DO and POSAT percentiles.Rdata"))
# Percentile file does not include the seasonal minimum

# Load the summer complete data
load(here("data","5.01 DR_Penn_Landing_Summer_Complete 2002-2022.Rdata"))
load(here("data","5. DR_Chester_Summer_Complete.Rdata"))

```

## Evaluating an unlikely scenario
Commenters noted that it was possible for POSAT to be less than the 66% criterion value for up to 24 days without violating the EPA's criteria. For this to happen, POSAT would have to be less than 66% for the period June 18 - June 30 (12 days in the Spawning and Larval Development Season) and July 1 - July 12 (12 days in the Juvenile Development Season).

Here we evaluate if it is likely for this scenario to occur. 

To begin, we need can compute the number of days that percent oxygen saturation was less than 66% in the period from June 19 and July 12 and relate it to the 10th percentile value of percent oxygen saturation in the _Juvenile Development_ season. The number of attaining days is bounded by 0 and 24, so it can be expressed as a fraction on the interval (0,1) and modeled using beta regression.

```{r}

# What is the highest value in each year from June 18 to July 12
seasonalMax <- rbind(
  drc_summer %>% mutate(Site="Chester"),
  drpl_summer %>% mutate(Site="Penn's Landing")) %>%
  mutate(doy.as.date = as.Date(paste0("1970-",format(date,"%m-%d")))) %>%  
  filter(between(doy.as.date,as.Date("1970-06-19"),as.Date("1970-07-12"))) %>% 
  group_by(Site,year) %>% 
  summarize(POSAT.max = max(posat.mean),
            POSAT.min = min(posat.mean))

numberAttains <- rbind(
  drc_summer %>% mutate(Site="Chester"),
  drpl_summer %>% mutate(Site="Penn's Landing")) %>%
  mutate(doy.as.date = as.Date(paste0("1970-",format(date,"%m-%d")))) %>%  
  filter(between(doy.as.date,as.Date("1970-06-19"),as.Date("1970-07-12"))) %>%
  mutate(attains = cut(posat.mean,breaks = c(0,66,1000),labels=c("not attain","attain"))) %>% 
  group_by(Site,year,attains) %>% 
  summarize(n=n()) %>% 
  pivot_wider(id_cols = c("Site","year"),names_from = "attains",values_from = "n",values_fill=0)

# Extract the POSAT 10th percentile from the percentiles variable 
p10 <- percentiles %>%
  filter(percentile==0.1,type=="Observed") %>% 
  mutate(p10=posat) %>% 
  select(Site,year,p10)

# Years the 10th percentile criterion was attained.
p10[p10$p10>66,]

# Joint them together
df <- left_join(seasonalMax,p10,by=c("Site","year")) %>% 
      left_join(.,numberAttains,by=c("Site","year")) %>% 
  ungroup() %>% 
  rename(not_attain = `not attain`) %>% 
  mutate(frac_attaining = (24 - not_attain)/24) 

summary(df)

beta_reg <- betareg(frac_attaining ~ p10,data=df %>% filter(Site=="Chester"))
summary(beta_reg)

tmp <- data.frame(p10=seq(40,70))
tmp$pred_frac <- predict(beta_reg,tmp)
tmp[tmp$p10==66,]
# 0.8334922*24  
# 20 days > 66% when P10 is 66%
  
ggplot()+
    geom_point(data=df %>% filter(Site=="Chester"),aes(x=p10,y=frac_attaining))+
    geom_vline(xintercept=66,linetype=2)+
    theme(panel.grid=element_blank())+
    geom_line(data=tmp,aes(x=as.numeric(p10),y=pred_frac))+
    labs(x="10th Percentile POSAT during Juvenile Development season",
         y="Fraction of POSAT >66% during June 19 - July 12")+
    ggtitle("Fraction Attaining 66% at Chester")

```
At Chester, we can predict that 20 of 24 days will have POSAT>66% if the 10th percentile of POSAT during the _Juvenile Development_ season is 66%. During 2002-2022, when the 10th percentile of POSAT was >66%, 8 days in 2015 had POSAT<66% and 1 day in 2013.  In 2016, 2017 and 2018, all the days during June 19 - July 12 had POSAT>66%.

```{r}

df[df$p10>66,]

```

We can fit the same model for Penn's Landing.

```{r}

beta_reg <- betareg(frac_attaining ~ p10,data=df %>% filter(Site=="Penn's Landing"))
summary(beta_reg)

tmp <- data.frame(p10=seq(40,70))
tmp$pred_frac <- predict(beta_reg,tmp)
tmp[tmp$p10==66,]
 0.9782295*24

ggplot()+
    geom_point(data=df %>% filter(Site=="Penn's Landing"),aes(x=p10,y=frac_attaining))+
    geom_vline(xintercept=66,linetype=2)+
    theme(panel.grid=element_blank())+
    geom_line(data=tmp,aes(x=as.numeric(p10),y=pred_frac))+
    labs(x="10th Percentile POSAT during Juvenile Development season",
         y="Fraction of POSAT >66% during June 19 - July 12")+
    ggtitle("Fraction Attaining 66% at Penn's Landing")

```


At Penn's Landing, when the 10th percentile of POSAT during the _Juvenile Development_ season is 66%, we can expect that POSAT will be less than 66% less than 1 day during June 19 - July 12. 

Based on this analysis, it can be concluded that the commenters scenario is extremely unlikely.
